html/semantics/embedded-content-0/media-elements/seeking/seek-to-currentTime.html

<!doctype html>
<title>seek to currentTime</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var v = document.createElement('video');
  v.src = getVideoURI('/media/movie_5');
  v.onloadedmetadata = t.step_func(function() {
    assert_greater_than(v.readyState, v.HAVE_NOTHING);
    assert_equals(v.seekable.start(0), 0);
    assert_greater_than(v.seekable.end(0), 0);
    assert_false(v.seeking);
    v.currentTime = v.currentTime;
    assert_true(v.seeking);
    var events = [];
    v.onseeking = v.ontimeupdate = v.onseeked = t.step_func(function(e) {
      events.push(e.type);
      assert_equals(v.seeking, e.type == 'seeking');
      if (e.type == 'seeked') {
        assert_array_equals(events, ['seeking', 'timeupdate', 'seeked']);
        t.done();
      }
    });
  });
});
</script>
